HW2 fix and renamed
