+++ app/src/androidTest/java/com/example/Calc/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.lesson1;
+package com.example.Calc;
 
 import android.content.Context;
