+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.lesson1">
+    package="com.example.Calc">
 
     <application
         android:allowBackup="true"
@@ -9,8 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Lesson1">
-        <activity android:name=".ActivityTwo"></activity>
-        <activity android:name=".MainActivity">
+        <activity android:name="com.example.Calc.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
