+++ app/src/main/java/com/example/Calc/MainActivity.java
@@ -1,14 +1,13 @@
-package com.example.lesson1;
+package com.example.Calc;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class ActivityTwo extends AppCompatActivity {
-
+public class MainActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_two);
+        setContentView(R.layout.activity_main);
     }
 }
