--- app/src/main/java/com/example/lesson1/MainActivity.java
@@ -1,58 +1,0 @@
-package com.example.lesson1;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.annotation.SuppressLint;
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.PersistableBundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.TextView;
-
-import java.text.BreakIterator;
-
-public class MainActivity extends AppCompatActivity implements View.OnClickListener {
-    private Integer arg1;
-    private Integer arg2;
-    Button btnActTwo;
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        btnActTwo = findViewById(R.id.btnActTwo);
-        btnActTwo.setOnClickListener(this);
-
-        EditText editText1 = findViewById(R.id.edit1);
-        EditText editText2 = findViewById(R.id.edit2);
-        Button button = findViewById(R.id.compare);
-        button.setOnClickListener(v -> {
-            TextView result = findViewById(R.id.result);
-
-            try {
-                arg1 = Integer.valueOf(editText1.getText().toString());
-                arg2 = Integer.valueOf(editText2.getText().toString());
-                if (arg1.equals(arg2)) {
-                    result.setText("Ravno");
-                } else {
-                    result.setText("Ne Ravno");
-                }
-            } catch (NumberFormatException e) {
-                result.setText("Введите число!");
-            }
-        });
-    }
-
-    @Override
-    public void onClick(View v) {
-        if (v.getId() == R.id.btnActTwo) {
-            Intent intent = new Intent(this, ActivityTwo.class);
-            startActivity(intent);
-        }
-    }
-}
